% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AbunTNIPlot}
\alias{AbunTNIPlot}
\title{Visualization of spots with tumor cell abundances and tumor-normal interface
(TNI) regions}
\usage{
AbunTNIPlot(
  abun,
  label,
  l_nshow,
  st = NULL,
  pos = NULL,
  size = 2.5,
  shape = 21,
  stroke = 1.2,
  title = NULL,
  legend.name = NULL,
  savefile = NULL,
  return = F,
  scale_y_reverse = F,
  scale_x_reverse = F,
  p.width = 10,
  p.height = 10,
  f.color = c("#0077b6", "lightyellow", "#c32f27"),
  limits = NULL,
  line_col = "black"
)
}
\arguments{
\item{abun}{Numeric vector of tumor cell abundances.}

\item{label}{labels of spots.}

\item{l_nshow}{labels represent nonTNI. Spots with these labels will not be
highlighted.}

\item{st}{Seurat object for ST data. If available, the coordinate information
should be included in the \code{metadata} as separate columns labeled 'x' and 'y'.}

\item{pos}{A data.frame with two columns named as 'x' and 'y', representing
the coordinate information; One of st and pos must be provided.}

\item{size}{Size of the spots.}

\item{shape}{A numerical value to set the shape of the TNI spots. Default: 21.}

\item{stroke}{A numerical value to set the stroke thickness.}

\item{title}{Title of plot.}

\item{legend.name}{Name of the legend.}

\item{savefile}{File name to save this plot. Only supporting pdf files.}

\item{return}{Logical indicating if the plot should be print or returned.}

\item{scale_y_reverse}{Logical indicating whether reversing the y position.}

\item{scale_x_reverse}{Logical indicating whether reversing the x position.}

\item{p.width}{The width of pdf file. The default value is 10.}

\item{p.height}{The height of pdf file. The default value is 10.}

\item{f.color}{Fill colors of spots.}

\item{limits}{Scaling limits. Any points outside these limits will not be
plotted.}

\item{line_col}{The color of the border of TNI spots.}
}
\description{
Visualization of spots with tumor cell abundances and tumor-normal interface
(TNI) regions
}
