% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpace.R
\name{BayesCluster}
\alias{BayesCluster}
\title{Title Domain annotation with BayesSpace}
\usage{
BayesCluster(
  st,
  assay = "Spatial",
  slot = "count",
  platfrom = "Visium",
  num_clu = NULL,
  qs = seq(3, 7),
  show_qplot = T,
  verbose = T,
  seed = 123,
  ...
)
}
\arguments{
\item{st}{Seurat object for ST data.}

\item{assay}{Name of assay.}

\item{slot}{Name of slot.}

\item{platfrom}{Spatial transcriptomic platform. Specify 'Visium' for hex
lattice geometry or 'ST' for square lattice geometry.}

\item{num_clu}{Number of clusters. If NULL, it will be selected
automatically.}

\item{qs}{A vector of continuous integers. Given the range to select the
number of cluster. Only used when \code{num.clu = NULL}. Default 3-7.}

\item{show_qplot}{Logical value. If TURE, plots the pseudo-log-likelihood
to help choose num_clu.}

\item{verbose}{Logical value. If TURE, print messages.}

\item{seed}{Set the seed.}

\item{...}{Other arguments passed on to \emph{spatialCluster}.}
}
\value{
Returns a modified \code{sce} with cluster assignments stored in
  \code{colData} under the name \code{spatial.cluster}.
}
\description{
Title Domain annotation with BayesSpace
}
