% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_interaction.R
\name{BuildLRExprssion}
\alias{BuildLRExprssion}
\title{Calculation of effective expression of ligand and receptor proteins.}
\usage{
BuildLRExprssion(
  Expr,
  LR_input,
  st_pos = NULL,
  do.dist = T,
  r.dist = 4,
  long.dist.method = "weighted",
  adjust = 2,
  na.rm = T,
  verbose = T
)
}
\arguments{
\item{Expr}{Expression data. Rows are genes and columns are spots.}

\item{LR_input}{A data.frame containing information of ligand and receptor
proteins.}

\item{st_pos}{A data.frame with coordinate information in two columns named
'x' and 'y'.}

\item{do.dist}{Logical value. If TURE, the expression of secreted proteins
will re-calculated using neighboring spots based on distance.}

\item{r.dist}{A numeric value. Distance to define the neighboring spots.
Default is 4.}

\item{long.dist.method}{Methods to integrate expression from spot and it's
neighbors. It should be one of 'weighted' and 'mean'.}

\item{adjust}{A numeric value controlling the weights of neighboring spots
with difference distance. Default is 2.}

\item{na.rm}{Logical value. If TURE, the proteins with no gene in \code{Expr}
will be removed.}

\item{verbose}{Logical value. If TURE, print messages.}
}
\value{
A data.frame of expression data. Rows are ligand and receptor
  proteins, and columns are spots.
}
\description{
Calculation of effective expression of ligand and receptor proteins.
}
