% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{BulidCGnet}
\alias{BulidCGnet}
\title{WGCNA analysis}
\usage{
BulidCGnet(
  data,
  outdir = NULL,
  power.seq = NULL,
  power = NULL,
  verbose = 3,
  ...
)
}
\arguments{
\item{data}{Expression data. Rows are selected genes, and columns are spots.}

\item{outdir}{Directory to save output files.}

\item{power.seq}{A vector of soft thresholding powers for which the scale
free topology fit indices are to be calculated.}

\item{power}{soft-thresholding power for network construction. If NULL, it
will be selected from \code{power.seq}.}

\item{verbose}{integer level of verbosity. Zero means silent, higher values
make the output progressively more and more verbose. Default is 3.}

\item{...}{Other arguments passed to \code{blockwiseModules} function in \pkg{WGCNA}.}
}
\value{
A list containing multiply components.
}
\description{
WGCNA analysis
}
