% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{CalCellCodis}
\alias{CalCellCodis}
\title{Calculation of cell co-distribution scores.}
\usage{
CalCellCodis(data, sort = T)
}
\arguments{
\item{data}{A matrix or data.frame containing abundances or proportions of
cell types. Rows are spots and columns are cell types.}

\item{sort}{Logical value indicating whether to sort the names of cells.}
}
\value{
A data.frame of co-distribution scores for all cell-cell pairs.
}
\description{
Calculation of cell co-distribution scores.
}
