% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{CalEntropy}
\alias{CalEntropy}
\title{Calculation of Shannon's Entropy}
\usage{
CalEntropy(x, base = 2)
}
\arguments{
\item{x}{a numeric probability vector.}

\item{base}{a character string specifying the logarithm unit. The default
value is 2.}
}
\value{
a numeric value representing Shannon's Entropy
}
\description{
Calculation of Shannon's Entropy
}
