% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{CalImmInfiltration}
\alias{CalImmInfiltration}
\title{Measuring the degree of immune infiltration using two measures: immune
enrichment and immune diversity}
\usage{
CalImmInfiltration(data, base = 2, min.prop = 0.05)
}
\arguments{
\item{data}{A matrix or data.frame containing proportions of
immune cells. Rows are spots and columns are cell types.}

\item{base}{base a character string specifying the logarithm unit. The
default value is 2.}

\item{min.prop}{Immune cells with a proportion less than this value will be set
to 0.}
}
\value{
A data.frame containing two columns: Imm.diversity and
  Imm.infiltration
}
\description{
Measuring the degree of immune infiltration using two measures: immune
enrichment and immune diversity
}
