% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_interaction.R
\name{CalLRIScore}
\alias{CalLRIScore}
\title{Calculation of ligand-receptor interaction (LRI) scores.}
\usage{
CalLRIScore(interaction_input, LR_expression, na.rm = T, p = NULL, verbose = T)
}
\arguments{
\item{interaction_input}{A data.frame containing information of
ligand-receptor interactions.}

\item{LR_expression}{A data.frame containing expression data of ligand and
receptor proteins.}

\item{na.rm}{Logical value. If TURE, LRIs with with all NA values will be
removed.}

\item{p}{The number of CPU cores used for parallel computing.}

\item{verbose}{Logical value. If TURE, print messages.}
}
\value{
A data.frame of LRI scores. Rows are LRIs, and columns are spots.
}
\description{
Calculation of ligand-receptor interaction (LRI) scores.
}
