% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLS_Score.R
\name{CalTLSfea}
\alias{CalTLSfea}
\title{Calculation of TLS scores. TLS features stored in}
\usage{
CalTLSfea(data, st_pos, cluster, ...)
}
\arguments{
\item{data}{A data.frame should contain three TLS features, including
co-local scores of "Plasma/B.cells_T.cells" and enrichment scores of two
signatures: "LC.50sig", "imprint.65sig".}

\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{cluster}{Cluster labels of spots.}

\item{...}{Arguments passed to other methods.}
}
\value{
Lists containing distinct forms of TLS features and the final TLS scores.
}
\description{
Calculation of TLS scores. TLS features stored in
}
\seealso{
[RegFeaEnrich]
}
