% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{CreateStfObj}
\alias{CreateStfObj}
\title{Create STFeature object}
\usage{
CreateStfObj(
  st,
  stf = NULL,
  assay = "SCT",
  slot = "data",
  cell.anno = NULL,
  norm = T,
  min.prop = 0,
  init.fea = c("Position", "CellAnno", "CellCodis"),
  verbose = T
)
}
\arguments{
\item{st}{Seurat object for ST data.}

\item{stf}{STFeature object. If available, overwrite it with specified
features.}

\item{assay}{Name of \code{assay} in \code{st}.}

\item{slot}{Name of \code{slot} in \code{st}.}

\item{cell.anno}{A data.frame containing annotation of cell types. Rows are
spots and columns are cell types. The values can be abundances or
proportions of cell types.}

\item{norm}{Logical value. If TRUE, the sum of abundances of each spot will
normalized to 1.}

\item{min.prop}{Cell types with normalized abundance below this value will be
set to 0. Default is 0.}

\item{init.fea}{Features to be initialized. It should be selected from
\code{c("Position", "CellCodis", "SpeGenes")}(Default).}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
A STFeature object.
}
\description{
Create STFeature object
}
