% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CuratedSig.lt}
\alias{CuratedSig.lt}
\title{Collected signatures from literatures.}
\format{
A list containing various gene sets.
}
\usage{
CuratedSig.lt
}
\description{
Collected signatures from literatures.
}
\concept{data}
\keyword{datasets}
