% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_interface.R
\name{DefineTNIregion}
\alias{DefineTNIregion}
\title{Identification of tumor-normal interface (TNI) regions}
\usage{
DefineTNIregion(
  x,
  st_pos,
  r.dist = 2,
  maxval = 0.08,
  minval = 0.03,
  boun.nei.n = 2,
  candi.step = 1,
  max.iter.num = 20,
  min.spots = 10,
  connect.dist = 2,
  verbose = T
)
}
\arguments{
\item{x}{A vector of TNI scores.}

\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{r.dist}{A numeric value. Distance to define the neighboring spots.
Default is 2.}

\item{maxval}{A numeric value. Spots with TNI scores above this value are
defiend as TNI spots.}

\item{minval}{A numeric value. Spots with TNI scores below this value are
defiend as non-TNI spots.}

\item{boun.nei.n}{A numeric value. The number of neighboring spots belong to
TNI.}

\item{candi.step}{An integer indicating the number of steps to add
\code{candi} spots when counting neighboring TNI spots. The higher the
value, the more contiguous the TNI area obtained. Default is 1.}

\item{max.iter.num}{The maximum number of times to update the spot labels.}

\item{min.spots}{A numeric value. Connected TNI regions with the number of
spots that fall below this value will be filtered out.}

\item{connect.dist}{Distance to define the spatially connected TNI regions.}

\item{verbose}{Logical value. If TURE, print messages.}
}
\value{
A vector of labels named with spots.
}
\description{
Identification of tumor-normal interface (TNI) regions
}
