% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{FactorEnrichAnalysis}
\alias{FactorEnrichAnalysis}
\title{Enrichment analysis for meta-genes of factors}
\usage{
FactorEnrichAnalysis(
  mg_vt,
  fun = "enrichGO",
  plot = T,
  show_num = 5,
  simplify = T,
  savefile = NULL,
  wrap_width = 50,
  p.width = 8,
  p.height = 10,
  ...
)
}
\arguments{
\item{mg_vt}{The vector of meta-genes.}

\item{fun}{One of "groupGO", "enrichGO", "enrichKEGG", "enrichDO" or
"enrichPathway".}

\item{plot}{A logical value indicating whether to plot.}

\item{show_num}{The number of terms to be displayed.}

\item{simplify}{A logical value indicating whether to simplify output from
compareCluster by removing redundancy of enriched GO terms.}

\item{savefile}{File name to save this plot (pdf).}

\item{wrap_width}{Positive integer giving target line width in characters. A
width less than or equal to 1 will put each word on its own line.}

\item{p.width}{The width of pdf file.}

\item{p.height}{The height of pdf file.}

\item{...}{Other arguments passed on to \code{compareCluster} in \pkg{clusterProfiler}.}
}
\value{
A compareClusterResult object.
}
\description{
Enrichment analysis for meta-genes of factors
}
