% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{FactorMetagenes}
\alias{FactorMetagenes}
\title{Meta-genes of each factors}
\usage{
FactorMetagenes(ref_W, top_num = 200)
}
\arguments{
\item{ref_W}{A data.frame contains the W matrix obtained from NMF analysis.}

\item{top_num}{Number of top genes.}
}
\value{
A vector of factors named with meta-genes.
}
\description{
Meta-genes of each factors
}
