% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneLen}
\alias{GeneLen}
\title{Gene length
#' @format A data.frame with one column: a numeric vector of gene length.
Rows are genes.}
\format{
An object of class \code{data.frame} with 57301 rows and 1 columns.
}
\usage{
GeneLen
}
\description{
Gene length
#' @format A data.frame with one column: a numeric vector of gene length.
Rows are genes.
}
\concept{data}
\keyword{datasets}
