% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{GetGsetSigScore}
\alias{GetGsetSigScore}
\title{Getting gene sets and accessing enrichment scores.}
\usage{
GetGsetSigScore(
  expr,
  stf = NULL,
  category = "CuratedSig",
  types = c("Immune", "Tumor"),
  subtype = NULL,
  species = "Homo sapiens",
  method = "AddModuleScore",
  scale = F,
  verbose = T
)
}
\arguments{
\item{expr}{Expression data. Rows are genes and Columns are spots.}

\item{stf}{STFeature object to store the enrichment scores.}

\item{category}{Category of gene sets. it should be one of "CuratedSig" and
"MSigDB".}

\item{types}{Types of gene sets. If \code{category} is "CuratedSig", it
should be one of "Immune" and "Tumor". If \code{category} is "MSigDB", the
types of gene sets can be checked with msigdb_types function.}

\item{subtype}{Subtypes of gene sets. For "Immune" type, it should be one of
"Immunoresponse", "PancancerCellAtlas", and "TLS; for "Tumor" type, it
should be one of "CancerMPs", "RecurTumorModules", "CancerSEA", and "Fges".}

\item{species}{Species name. Default is Mus musculus.}

\item{method}{Method for enrichment analysis, passed on to \code{GsetScore}}

\item{scale}{Logical value. If TRUE, scaling and centering the scores for all
spots.}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
If \code{stf} is NULL, return a data with enrichment scores; else,
  return stf with \code{GsetSig} slot added.
}
\description{
SpaTME enables exploration of signatures or gene sets collected
  from literatures and databases.
}
