% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_interface.R
\name{GroupTNItypes}
\alias{GroupTNItypes}
\title{Grouping TNI spots into distinct types.}
\usage{
GroupTNItypes(
  TNI_pos,
  cluster,
  seed = 123,
  type_n = NULL,
  cluster_method = "pam",
  nei.dist = 4,
  connect.dist = 4,
  min.type.spots = 0,
  disconnect.spots = 3,
  verbose = T
)
}
\arguments{
\item{TNI_pos}{Position of TNI spots. A data.frame with coordinate
information in two columns.}

\item{cluster}{A vector of cluster labels of TNI spots.}

\item{seed}{Seed for clustering spots.}

\item{type_n}{Number of types to cluster.}

\item{cluster_method}{Clustering method, one of "kmeans", "hclust", and
"pam".}

\item{nei.dist}{A numeric value. Distance to define the neighboring spots.
Default is 4.}

\item{connect.dist}{A numeric value. Distance to define the spatially
connected TNI regions within each type. Default is 4.}

\item{min.type.spots}{A numeric value. Types with number of spots less than this value will be filtered.}

\item{disconnect.spots}{A numeric value. For each TNI type, connected TNI regions with the number of
spots that fall below this value will be filtered out. Default is 3.}

\item{verbose}{Logical value. If TURE, print messages.}
}
\value{
A vector of TNI types named with spots.
}
\description{
Grouping TNI spots into distinct types.
}
