% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{GsetScore}
\alias{GsetScore}
\title{Enrichment scores of gene sets.}
\usage{
GsetScore(
  expr,
  geneset,
  method = "AddModuleScore",
  gsva.method = "ssgsea",
  scale = F,
  verbose = T
)
}
\arguments{
\item{expr}{Expression data. Rows are genes and Columns are spots.}

\item{geneset}{A list of vectors of features; each entry should be named with
feature.}

\item{method}{Method for enrichment analysis.  Available options are:
\itemize{
  \item{"AddModuleScore"}: Calculate the average expression levels of each
  gene set for each spot, subtracted by the aggregated expression of control
  feature sets. This method is conducted using \pkg{Seurat}.
  \item{"UCell"}: Calculate module enrichment scores using \pkg{UCell}.
  \item{"AUCell"}: Calculates the 'AUC' for each gene-set in each spot uising
  functions in \pkg{AUCell}.
  \item{"gsva"}: Estimates enrichment scores
  using \pkg{GSVA}.
}}

\item{gsva.method}{Method employed in the estimation of \pkg{GSVA}. default
is \code{ssgsva}.}

\item{scale}{Logical value. If TRUE, scaling and centering the scores for all spots.}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
A data.frame containing the enrichment scores.
}
\description{
Enrichment scores of gene sets.
}
