% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{HPhenoAsso}
\alias{HPhenoAsso}
\title{Confirming phenotype-related NMF factors}
\usage{
HPhenoAsso(nmfobj, phenotype, method, ...)
}
\arguments{
\item{nmfobj}{An object of class NMFfit.}

\item{phenotype}{A numeric or character vector representing the annotation of spots or
phenotype of samples.}

\item{...}{Other arguments passed to \code{PhenoAssoFeatures}.}
}
\value{
A character vector represent association between phenotype and factors.
}
\description{
Confirming phenotype-related NMF factors
}
