% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Immunomodulators}
\alias{Immunomodulators}
\title{Immune-related genes}
\format{
A vector of genes, which play a critical role in cancer immunotherapy.
}
\usage{
Immunomodulators
}
\description{
Immune-related genes
}
\concept{data}
\keyword{datasets}
