% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interaction_input}
\alias{interaction_input}
\title{Ligand-receptor interaction pairs.}
\format{
A data.frame.
}
\usage{
interaction_input
}
\description{
Ligand-receptor interaction pairs.
}
\concept{data}
\keyword{datasets}
