% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{LabelMapcolor}
\alias{LabelMapcolor}
\title{Colors of labels.}
\usage{
LabelMapcolor(
  labels,
  na.col = "grey",
  assgin.col = NULL,
  colors = NULL,
  seed = NULL,
  random = F
)
}
\arguments{
\item{labels}{A vector of labels to set colors. It should be a character or
factor vector.}

\item{na.col}{Color for NA values.}

\item{assgin.col}{A vector of labels assigned colors by users. Colors and
labels included in this vector will not be assigned again.}

\item{colors}{Characters vector of colors assigned to labels. If NULL, the
default colors are used.}

\item{seed}{An integer to set the seed.}

\item{random}{Logical indicating whether colors are assigned to labels
sequentially or randomly.}
}
\value{
A vector of colors named with labels
}
\description{
Colors of labels.
}
\examples{
x <- c("A", "B", "C", "D")
cols <- LabelMapcolor(x, assgin.col = c(A = "red"))
}
