% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{msigdb_types}
\alias{msigdb_types}
\title{Gene set types in MSigDB.}
\usage{
msigdb_types()
}
\value{
A character vector represents types of gene sets in "MSigDB" database.
}
\description{
Gene set types in MSigDB.
}
\examples{
msigdb_types()
}
