% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{NMF_bulk_input}
\alias{NMF_bulk_input}
\title{Preparing input data for NMF analysis.}
\usage{
NMF_bulk_input(
  data,
  totpm = F,
  genelen = NULL,
  dolog = F,
  pt_gene_exp = 0.1,
  base_gene_exp = 0,
  pseudogene.filt = F
)
}
\arguments{
\item{data}{Expression data. Rows are genes and columns are samples.}

\item{totpm}{Logical value. If TURE, convert the counts to TPM values.}

\item{genelen}{A data.frame containing the gene length information. Only used when \code{totpm = TURE}.}

\item{dolog}{Logical value. If TURE, conduct log2 transformation.}

\item{pt_gene_exp}{Only keep genes that are detected in a minimum fraction (pt_gene_exp) of samples.}

\item{base_gene_exp}{Only genes with expression above this value are considered to be expressed.}

\item{pseudogene.filt}{Logical value indicating whether to filt pseudogene.}
}
\value{
A data.frame with pre-processed expression profiles.
}
\description{
Preparing input data for NMF analysis.
}
