% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_features.R
\name{Norm01}
\alias{Norm01}
\title{Normalizing numeric data between 0 and 1.}
\usage{
Norm01(data, constant.v = 0)
}
\arguments{
\item{data}{A numeric vector, matrix or data.frame.}

\item{constant.v}{Constent value when there is no varience in \code{data}.
Default is 0.}
}
\value{
Normalized data.
}
\description{
Normalizing numeric data between 0 and 1.
}
