% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{PhenoAssoFeatures}
\alias{PhenoAssoFeatures}
\title{Identification of phenotype or annotation associated features.}
\usage{
PhenoAssoFeatures(
  data,
  phenotype,
  method,
  family = NULL,
  cor.method = "pearson",
  setNaN = 1,
  p.adj = F,
  deseq2_contrast = NULL,
  adj.method = "BH",
  p.cut = 0.05,
  p.adj.cut = 0.05,
  auc.power = 0.4,
  cox.test.cut = 0.05,
  logFC = 0.5,
  verbose = T
)
}
\arguments{
\item{data}{A data matrix or data frame; col are samples and row are features.}

\item{phenotype}{A numeric or character vector representing the annotation of spots or
phenotype of samples.}

\item{method}{Method to use. Available options are:
\itemize{
 \item{"cox"} ：Fits a Cox proportional hazards regression model. This is used
 when the phenotype is survival time of patients.
 \item{"wilcox"} : Identifies differential features between two
 groups of spots or samples using a Wilcoxon Rank Sum test.
 \item{"glm"} Fits generalized linear models.
 \item{"auc"} A metric obtained from ROC analysis.
 \item{"deseq2"} Identifies differential features between two groups using
 \pkg{DESeq2}. Only integer values are supported.
 \item{"cor"} Test for association/correlation between paired samples/spots. It is used for continuous variables.
 }}

\item{family}{a character string naming a family function when \code{method} is "glm".}

\item{cor.method}{Method used for correlation coefficient. One of "pearson",
"kendall", or "spearman", can be abbreviated.}

\item{p.adj}{Logical value indicating whether to adjust the p values.}

\item{deseq2_contrast}{Argument specifies the groups for comparison. It is
passed to \code{results} function in \pkg{DESeq2}.}

\item{adj.method}{Method to adjust p values. It should be one of
\code{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr",
"none")}. Default is "BH".}

\item{p.cut}{A cutoff of p values. It is ignored if \code{p.adj = TRUE}. Default is 0.05.}

\item{p.adj.cut}{A cutoff of adjusted p values. It is ignored if \code{p.adj = FALSE}. Default is 0.05.}

\item{auc.power}{A cutoff of absolute value of auc power. It is used when
\code{method = "auc"}. The auc power is defined as (AUC-0.5) * 2. Default is 0.4.}

\item{cox.test.cut}{A cutoff of p values in proportional hazards assumption
test. Features with p value above this value are kept. Default is 0.05.}

\item{logFC}{A cutoff of log(fold change). Default is 0.5}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
A data.frame with features as rows, and associated statistics as
  columns (p-values, AUC power, etc.), depending on the methods.
}
\description{
To explore the association between features and specific phenotypes of
patients or annotations of spots, this function utilizes various methods that
can be selected based on the type of the target variable.
}
\details{
Except for DESeq2 method, this calculation is performed separately for each feature in the given data.
}
