% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotCellSpot}
\alias{PlotCellSpot}
\title{Spatial plot of cell distribution}
\usage{
PlotCellSpot(
  decon_mtrx,
  st_pos,
  tarCells = NULL,
  savefile = NULL,
  separate = T,
  f.color = c("#0077b6", "lightyellow", "#c32f27"),
  e.color = "black",
  pie_scale = 0.7,
  pie_color = NULL,
  size = 5,
  size.auto = F,
  size.factor = 0.36,
  numCol = 4,
  p.width = 10,
  p.height = 10,
  scale_y_reverse = F,
  scale_x_reverse = F
)
}
\arguments{
\item{decon_mtrx}{A numeric matrix or data.frame containing cell abundances or
percents for each spot. Rows are spots and columns are cell types.}

\item{st_pos}{A data.frame with coordinate information in two columns named 'x'
and 'y'.}

\item{tarCells}{Cell types selected to display. If NULL, all cell types in
decon_mtrx are used.}

\item{savefile}{File name to save this plot. Only supporting pdf files.}

\item{separate}{Logical value. If TRUE, the default, each cell types will
plot separately. If FALSE, the distribution of all selected cell types are
showed with pie charts for each spot.}

\item{f.color}{Fill colors of spots.}

\item{e.color}{Edge color of pie.}

\item{pie_scale}{Amount to scale the pie size.}

\item{pie_color}{A character vector. Colors of cell types in pie.}

\item{size}{Size of spots when separate is TRUE.}

\item{size.auto}{Logical indicating if adjusting the size automatically based
on sample size. It is useful when drawing multiple samples.}

\item{size.factor}{A numeric value, scale the size of spots when \code{size.auto = TRUE}.}

\item{numCol}{The number of columns in one pdf file if \code{separate = TRUE}.}

\item{p.width}{The width of pdf file. The default value is 10.}

\item{p.height}{The height of pdf file. The default value is 10.}

\item{scale_y_reverse}{Logical indicating whether reversing the y position.}

\item{scale_x_reverse}{Logical indicating whether reversing the x position.}
}
\description{
Spatial plot of cell distribution
}
