% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{PredNMFinST}
\alias{PredNMFinST}
\title{Prediction of factors in ST data}
\usage{
PredNMFinST(
  st,
  W,
  assay = "SCT",
  slot = "data",
  plotfile = NULL,
  numCol = 1,
  verbose = T,
  ...
)
}
\arguments{
\item{st}{a seurat object for ST data.}

\item{W}{The W matrix in NMF analysis.}

\item{assay}{Name of \code{assay} in \code{st}.}

\item{slot}{Name of \code{slot} in \code{st}.}

\item{plotfile}{File name to save this plot. Only supporting pdf files.}

\item{numCol}{Number of columns in the plot grid.}

\item{verbose}{Logical value. If TURE, print messages.}

\item{...}{Other arguments passed to \code{SpotVisualize}.}
}
\value{
A NMFfit object with predicted new H matrix.
}
\description{
Prediction of factors in ST data
}
