% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{RunNMFtest}
\alias{RunNMFtest}
\title{NMF analysis to identify phenotype-related factors}
\usage{
RunNMFtest(
  expr,
  rank,
  nrun = 30,
  seed = 123456,
  min_cophenetic = 0.95,
  name_prefix = "Factor",
  return.all = F,
  verbose = T
)
}
\arguments{
\item{expr}{Expression data. Rows are phenotype-related genes and columns are
samples.}

\item{rank}{A single numeric value or a numeric vector. Number of factors.}

\item{nrun}{Number of runs to perform. Default is 30.}

\item{seed}{a single numeric that is used to seed the random number
generator, before generating a random starting point. Default is 123456.}

\item{min_cophenetic}{Minimum cophenetic coefficient. If the rank is a numeric
vector, the cophenetic coefficient is used to measure the stability of the
clusters obtained. The selected number of ranks should be above this value.
Defualt is 0.95.}

\item{name_prefix}{A character specified the prefix name of factors.}

\item{return.all}{Logical value}

\item{verbose}{Logical value. If TURE, print messages.}
}
\value{
If the parameter return.all is TURE and the rank is a numeric vector,
  return a list containing all the NMFfit objects of each rank and their
  cophenetic coefficients; else, return an object of class NMFfit.
}
\description{
NMF analysis to identify phenotype-related factors
}
