% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{SePreprocess}
\alias{SePreprocess}
\title{Title Preprocessing of ST data with SpaTME}
\usage{
SePreprocess(
  se,
  assay = "RNA",
  mt.qc = 30,
  norm.SCT = TRUE,
  variable.features.n = 3000,
  dims = 1:20,
  cluster.resolution = 0.8,
  verbose = F
)
}
\arguments{
\item{se}{A Seurat object of ST data to preprocess.}

\item{assay}{name of assay to use.}

\item{mt.qc}{Spots with percentage of reads that map to the mitochondrial
genome more than this value will be filtered. If NULL, no filtering.}

\item{norm.SCT}{Logical value. If TRUE, normalize data with SCTransform.
If FALSE, running NormalizeData, FindVariableFeatures, and ScaleData
Sequentially.}

\item{variable.features.n}{Number of features to select as top variable features.}

\item{dims}{Dimensions of reduction to use.}

\item{cluster.resolution}{Value of the resolution parameter.}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
A Seurat object with preprocessing results (like PCA, clustering, and UMAP).
}
\description{
Title Preprocessing of ST data with SpaTME
}
