% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpotVisualize}
\alias{SpotVisualize}
\title{Visualization of ST spots with features}
\usage{
SpotVisualize(
  st = NULL,
  pos = NULL,
  meta = NULL,
  feature = NULL,
  size = 2.5,
  savefile = NULL,
  return = F,
  cha.col = NULL,
  title = "NULL",
  legend.name = NULL,
  p.width = 10,
  p.height = 10,
  size.factor = 0.02,
  f.color = c("#0077b6", "lightyellow", "#c32f27"),
  scale_y_reverse = F,
  scale_x_reverse = F,
  limits = NULL,
  na.col = "#ced4da",
  ...
)
}
\arguments{
\item{st}{Seurat object for ST data. If available, the coordinate information
should be included in the \code{metadata} as separate columns labeled 'x'
and 'y'.}

\item{pos}{A data.frame with two columns named as 'x' and 'y', representing
the coordinate information; One of st and pos must be provided.}

\item{meta}{A vector of the feature to visualize. If st is available, it can
be one of the feature name in the metadata.}

\item{feature}{A character of gene name in \code{data} slot of st. Provide
either \code{meta} or \code{feature}, not both.}

\item{size}{Size of the spots.}

\item{savefile}{File name to save this plot. Only supporting pdf files.}

\item{return}{Logical indicating if the plot should be print or returned.}

\item{cha.col}{Colors of meta if the meta is a character vector. By default,
ggplot2 assigns colors.}

\item{title}{Title of plot.}

\item{legend.name}{Name of the legend.}

\item{p.width}{The width of pdf file. The default value is 10.}

\item{p.height}{The height of pdf file. The default value is 10.}

\item{size.factor}{A numeric value, scale the size of spots based on the
p.width and p.height. It will be ignored if size is set. The default value
is 0.02.}

\item{f.color}{Fill colors of spots.}

\item{scale_y_reverse}{Logical indicating whether reversing the y position.}

\item{scale_x_reverse}{Logical indicating whether reversing the x position.}

\item{limits}{Scaling limits. Any points outside these limits will not be
plotted.}

\item{na.col}{Color for NA values.}

\item{...}{Other arguments passed on to \code{theme} in \pkg{ggplot}.}
}
\value{
A ggplot object if \code{return = TRUE}.
}
\description{
Visualization of ST spots with features
}
