% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{STcoordCheck}
\alias{STcoordCheck}
\title{Coordinate check for ST data}
\usage{
STcoordCheck(
  st,
  reset = F,
  platform = "Visium",
  hexagon.correct = F,
  hexagon.direct = c("horizontal", "vertical"),
  verbose = T
)
}
\arguments{
\item{st}{Seurat object for ST data. The coordinate information will be
checked using the \code{metadata}, \code{images}, or spot names.}

\item{reset}{Logical indicating whether the existing 'x' and 'y' columns in
\code{metadata} should be reset.}

\item{platform}{Spatial transcriptomic platform. Specify 'Visium' for hex
lattice geometry or 'ST' for square lattice geometry.}

\item{hexagon.correct}{Logical value, indicating whether to correct the
coordinate position so that it conforms to the hexagon. Used only when
\code{platform == "visium"}.}

\item{hexagon.direct}{Direction of hexagon. Available options are:
\itemize{
 \item{"horizontal"} : spots to left and right, two above, two below.
 \item{"vertical"} : spots to top and bottom , two left and two right.
 }}

\item{verbose}{Logical value. If TRUE, print messages.}
}
\value{
Seurat object of ST data with checked coordinates in \code{metadata} named
  as x' and 'y'.
}
\description{
Coordinate check for ST data
}
