% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_program_NMF.R
\name{std.divergence.update.h}
\alias{std.divergence.update.h}
\title{NMF Multiplicative Updates for Kullback-Leibler Divergence}
\usage{
std.divergence.update.h(v, w, h, nbterms = 0L, ncterms = 0L, copy = TRUE)
}
\arguments{
\item{v}{target matrix.}

\item{w}{current basis matrix.}

\item{h}{current coefficient matrix.}

\item{nbterms}{number of fixed basis terms.}

\item{ncterms}{number of fixed coefficient terms.}

\item{copy}{logical that indicates if the update should be made on the original
matrix directly (\code{FALSE}) or on a copy (\code{TRUE} - default).}
}
\value{
a matrix of the same dimension as the input matrix to update.
}
\description{
NMF Multiplicative Updates for Kullback-Leibler Divergence
}
