% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_interface.R
\name{TNIscore}
\alias{TNIscore}
\title{Tumor-normal interface (TNI) scores}
\usage{
TNIscore(ES, st_pos, cluster = NULL, r.dist = 2, norm01 = F)
}
\arguments{
\item{ES}{Tumor cell abundances or other features representing the enrichment
of tumor cells.}

\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{cluster}{Cluster labels of spots.}

\item{r.dist}{A numeric value. Distance to define the neighboring spots.
Default is 2.}

\item{norm01}{Logical value indicating whether normalizing the TNI scores
within smaple to range 0-1.}
}
\value{
A vector of TNI scores.
}
\description{
Tumor-normal interface (TNI) scores
}
