% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{unit}
\alias{unit}
\alias{UnitEnrich}
\alias{RegFeaEnrich}
\title{Calculating enriched value in unit region.}
\usage{
UnitEnrich(
  data,
  d.lt,
  method = "mean",
  weight = NULL,
  layer.method = "mean",
  verbose = F
)

RegFeaEnrich(
  data,
  st_pos,
  cluster = NULL,
  r.dist = 2,
  method = "weighted",
  layer.method = "mean",
  adjust = 2,
  verbose = T
)
}
\arguments{
\item{data}{Data containing features of spots.}

\item{d.lt}{A lists of spots grouped by distance.}

\item{method}{Methods to aggregate values from neighbor layers.}

\item{weight}{Weights for each layer.}

\item{layer.method}{Methods to aggregate values for each neighbor layer.}

\item{verbose}{Logical value. If TURE, print messages.}

\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{cluster}{Clusters of spots.}

\item{r.dist}{A numeric value. Distance to define the neighboring spots.}

\item{adjust}{Adjustment of weights.}
}
\value{
Integrated value of feature in this unit.

Integrated values of features for each unit.
}
\description{
For each spot, the unit is defined as the central spot and its
  neighboring spots up to n-layers. Features within this unit could be
  integrated.
}
